fix(playground-e2e): correct relative import paths after spec move
The specs were moved from e2e/cypress/e2e/playground/<spec>.cy.ts (3
levels deep under cypress/) to packages/untp-playground/e2e/cypress/e2e/<spec>.cy.ts
(2 levels deep). The unchanged `../../support/config` import resolved to
the wrong directory; it should be `../support/config` at the new depth.

Verified locally with all 5 specs against `docker compose --profile playground`:
26 passing.
