+++ packages/untp-playground/e2e/cypress/e2e/download-report.cy.ts
@@ -1,4 +1,4 @@
-import { config } from '../../support/config';
+import { config } from '../support/config';
 
 describe('Download Report', () => {
   beforeEach(() => {
