+++ packages/untp-playground/e2e/cypress/e2e/report-generation.cy.ts
@@ -1,5 +1,5 @@
 import { TestCaseStatus } from '../../../constants';
-import { config } from '../../support/config';
+import { config } from '../support/config';
 
 describe('Report Generation', () => {
   beforeEach(() => {
