+++ packages/untp-playground/e2e/cypress/e2e/upload-invalid-credential.cy.ts
@@ -1,4 +1,4 @@
-import { config } from '../../support/config';
+import { config } from '../support/config';
 
 describe('Display Error Messages', () => {
   beforeEach(() => {
