+++ packages/untp-playground/e2e/cypress/e2e/vcdm-validation.cy.ts
@@ -1,5 +1,5 @@
 import { VCDM_CONTEXT_URLS } from '../../../constants';
-import { config } from '../../support/config';
+import { config } from '../support/config';
 
 describe('VCDM Schema Validation', () => {
   beforeEach(() => {
